absolute reference to List/ methods for String/
Without this, these functions are actually non-terminating, as they
just reference themselves.

There is an underlying bug here that still needs to be fixed, where
functions that are imported with the `use` syntax auto-fill erased
arguments. For example,

  use List/length

  length
  - s: (List Char)
  : Nat

  length s

will type-check, but

  length
  - s: (List Char)
  : Nat

  List/length s

won't, as `List/length` has an erased argument. This is a temporary
fix as the stringifier will preserve type-checking with these new
programs.