fix definition being shadowed by uses
if you attempt to define a term that has the same name as something that
is imported, it was previously given an incorrect name. speciifically,
it was given the absolute name of the used term. for example,

use List/{is_empty}

is_empty (str: String) = 0

would previously desugar as

List/is_empty (str: String) = 0

which is incorrect.